backend/requirements.txt
# Core dependencies
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
python-dotenv==1.0.0
python-multipart==0.0.6

# Authentication & Security
PyJWT==2.6.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Database
sqlalchemy==1.4.46
alembic==1.10.2
aiosqlite==0.18.0

# AI/ML
openai==0.27.8
numpy==1.24.3

# Twilio Integration
twilio==7.16.1

# LiveKit Integration
livekit-server-sdk>=0.9.0,<1.0.0  # Server SDK for LiveKit
grpcio>=1.54.0  # Required for gRPC
protobuf>=4.23.0  # Required for gRPC

# WebRTC
aiohttp==3.8.4
websockets==11.0.3

# Local modules
transcripts/  # Local module

# Development Dependencies (optional)
pytest==7.4.2
pytest-asyncio==0.21.1
black==23.9.1
isort==5.12.0
mypy>=1.5.1
flake8>=6.1.0

# Development
python-jose[cryptography]>=3.3.0  # JWT token handling
python-multipart>=0.0.6  # For file uploads

# Utilities
python-dateutil>=2.8.2
python-slugify>=8.0.1
email-validator>=2.0.0

# Async
anyio>=3.7.1
asgiref>=3.7.2

# CORS
starlette>=0.31.0

# Logging
structlog>=23.1.0
